chore(tools): keep one copy of the anti-slop lint rules
The plugin was committed twice, byte for byte, under tools/oxlint and as
the install-anti-slop skill's asset tree. Keep tools/oxlint, the copy the
skill's installer refuses to overwrite, and drop the skill with its lock
entry.

In the survivor, drop the rules that duplicate the compiler
(no-unknown-type-aliases, the weaker copy of no-unsafe-dictionary-type;
no-widen-then-assert, subsumed by no-known-value-widening), the rule that
targets a test framework this repository does not use
(no-module-mocking), the rule that bans a single English word
(no-shape-in-symbol-names), and the never-registered effect/ tree. Merge
no-reflect-apply and no-reflect-get into no-reflect, and
no-object-parameters into no-unknown-parameters, so one rule reports each
unparsed function input. Delete the type-guard escape hatch
no-runtime-typeof never had enabled, and the unreferenced
isPopulatedObjectExpression and WideningTargetKind.
