+++ skills-lock.json
@@ -1,11 +1,4 @@
 {
   "version": 1,
-  "skills": {
-    "install-anti-slop": {
-      "source": "dmmulroy/anti-slop",
-      "sourceType": "github",
-      "skillPath": "skills/install-anti-slop/SKILL.md",
-      "computedHash": "b903a7ac2c823d473a4b0046834fccbab0bc75ee4fda8be4f81efa895d475b01"
-    }
-  }
+  "skills": {}
 }
