--- tools/oxlint/anti-slop/effect/index.ts
@@ -1,13 +1,0 @@
-import { eslintCompatPlugin } from "@oxlint/plugins";
-
-import { noServiceConstructorImportsRule } from "./rules/no-service-constructor-imports.ts";
-
-/** Opt-in Oxlint rules for Effect service and Layer architecture. */
-const antiSlopEffectPlugin = eslintCompatPlugin({
-  meta: { name: "anti-slop-effect" },
-  rules: {
-    "no-service-constructor-imports": noServiceConstructorImportsRule,
-  },
-});
-
-export default antiSlopEffectPlugin;
