+++ tools/oxlint/anti-slop/rules/no-runtime-typeof.ts
@@ -1,26 +1,4 @@
-import type { ESTree } from "@oxlint/plugins";
 import { defineRule } from "@oxlint/plugins";
-
-type RuntimeFunction = ESTree.ArrowFunctionExpression | ESTree.Function;
-
-function isRuntimeFunction(node: ESTree.Node): node is RuntimeFunction {
-  return (
-    node.type === "ArrowFunctionExpression" ||
-    node.type === "FunctionDeclaration" ||
-    node.type === "FunctionExpression"
-  );
-}
-
-function isInsideTypeGuard(node: ESTree.Node): boolean {
-  let current: ESTree.Node | null = node.parent;
-  while (current !== null && current.type !== "Program") {
-    if (isRuntimeFunction(current)) {
-      return current.returnType?.typeAnnotation.type === "TSTypePredicate";
-    }
-    current = current.parent;
-  }
-  return false;
-}
 
 /** Disallow runtime typeof checks that narrow unparsed values instead of decoding them. */
 export const noRuntimeTypeofRule = defineRule({
@@ -34,27 +12,11 @@ export const noRuntimeTypeofRule = defineRule({
       runtimeTypeof:
         "A `typeof` check narrows a representation without establishing its contract. Parse input at its I/O boundary, then branch on the domain value.",
     },
-    schema: [
-      {
-        type: "object",
-        properties: {
-          allowInTypeGuards: { type: "boolean" },
-        },
-        additionalProperties: false,
-      },
-    ],
-    defaultOptions: [{ allowInTypeGuards: false }],
   },
   createOnce(context) {
     return {
       UnaryExpression(node) {
-        const option = context.options?.[0];
-        const allowInTypeGuards =
-          typeof option === "object" &&
-          option !== null &&
-          !Array.isArray(option) &&
-          option.allowInTypeGuards === true;
-        if (node.operator === "typeof" && (!allowInTypeGuards || !isInsideTypeGuard(node))) {
+        if (node.operator === "typeof") {
           context.report({ node, messageId: "runtimeTypeof" });
         }
       },
