+++ tools/oxlint/anti-slop/shared/dictionary-types.ts
@@ -24,15 +24,13 @@ export type UnsafeDictionary = {
   readonly unsafeValue: "any" | "empty-object" | "object" | "union" | "unknown";
 };
 
-export type WideningTargetKind =
-  | "anonymous object"
-  | "generic container"
-  | "object"
-  | "open dictionary"
-  | "unknown";
-
 export type WideningTarget = {
-  readonly kind: WideningTargetKind;
+  readonly kind:
+    | "anonymous object"
+    | "generic container"
+    | "object"
+    | "open dictionary"
+    | "unknown";
 };
 
 export type TypeEnvironment = {
@@ -454,19 +452,6 @@ function classifyAliasBroadTarget(
     nextSubstitutions,
     nextResolving,
   );
-}
-
-export function isPopulatedObjectExpression(expression: ESTree.Expression): boolean {
-  let current = expression;
-  while (
-    current.type === "ParenthesizedExpression" ||
-    current.type === "TSAsExpression" ||
-    current.type === "TSTypeAssertion" ||
-    current.type === "TSNonNullExpression"
-  ) {
-    current = current.expression;
-  }
-  return current.type === "ObjectExpression" && current.properties.length > 0;
 }
 
 export function isKnownEvidenceExpression(expression: ESTree.Expression): boolean {
